arc_primitives_solver/requirements.txt
# Primitives-based ARC-AGI-2 Solver
#
# Dependencies for new solver (reuses arc_solver's venv)

# Core
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pydantic>=2.0.0

# API client
httpx>=0.24.0
tenacity>=8.2.0

# Environment
python-dotenv>=1.0.0
